+++ CHANGELOG.md
@@ -36,6 +36,7 @@ All notable changes to this project will be documented in this file.
 - `live_response/modifiers/disable_ftrace.yaml`: Added modifier to disable ftrace to prevent syscall hooking by LKM rootkits [linux]. (by [mnrkbys](https://github.com/mnrkbys))
 - `live_response/packages/slackpkg.yaml`: Moved to `packages/slackpkg.yaml`.
 - `live_response/process/procfs_information.yaml`: Added collection of /proc/modules [aix, freebsd, linux, netbsd, netscaler, solaris]. (by [SolitudePy](https://github.com/SolitudePy))
+- `live_response/system/getcap.yaml`: Moved to `system/getcap.yaml`. Updated to skip non-local file systems [linux]. ([#375](https://github.com/tclahr/uac/issues/375))
 - `live_response/system/group_name_unknown_files.yaml`: Moved to `system/group_name_unknown_files.yaml`.
 - `live_response/system/hidden_directories.yaml`: Moved to `system/hidden_directories.yaml`.
 - `live_response/system/hidden_files.yaml`: Moved to `system/hidden_files.yaml`.